requirements.txt
fastapi==0.116.1
uvicorn==0.34.3
requests==2.31.0
python-dotenv==1.1.0
supabase==2.15.1
openai==1.92.2
google-generativeai==0.8.5
geopandas==1.1.0
shapely==2.1.1
folium==0.20.0
plotly==6.3.1
numpy==2.2.6
pandas==2.2.3
pyproj==3.7.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
python-docx==1.2.0
reportlab==4.4.3
python-multipart==0.0.9
pypdf
rasterio==1.4.3
qrcode[pil]
sendgrid==6.11.0
email-validator
pypandoc_binary
langchain>=0.2.0
langchain-openai>=0.1.7
langchain-core>=0.2.0
openai>=1.30.0
tiktoken>=0.7.0
pdf2image==1.17.0
mistralai==1.2.4
mammoth
